docs: explain how to replace editor content
