+++ packages/storybook-react/stories/react/replace-content-and-clear-history.tsx
@@ -1,0 +1,40 @@
+import { useCallback } from 'react';
+import { Remirror, ThemeProvider, useRemirror } from '@remirror/react';
+
+const DOC = {
+  type: 'doc',
+  content: [
+    {
+      type: 'paragraph',
+      content: [
+        {
+          type: 'text',
+          text: 'New content',
+        },
+      ],
+    },
+  ],
+};
+
+const ReplaceContentAndClearHistory = (): JSX.Element => {
+  const { manager, state, onChange } = useRemirror({
+    extensions: () => [],
+    content: '<p>Original content</p>',
+    stringHandler: 'html',
+  });
+
+  const handleClick = useCallback(() => {
+    // Clear out old state when setting data from outside
+    // This prevents e.g. the user from using CTRL-Z to go back to the old state
+    manager.view.updateState(manager.createState({ content: DOC }));
+  }, [manager]);
+
+  return (
+    <ThemeProvider>
+      <Remirror manager={manager} state={state} onChange={onChange} />
+      <button onClick={handleClick}>Replace content</button>
+    </ThemeProvider>
+  );
+};
+
+export default ReplaceContentAndClearHistory;
